requirements.txt
matplotlib==3.8.1
numba==0.58.1
numpy==1.26.2
Pillow==10.1.0
scikit_learn==1.3.2
scipy==1.11.3
torch==2.1.0
wandb==0.16.0
